+++ Source/Krypton Components/Krypton.Toolkit/General/GlobalStrings.cs
@@ -41,6 +41,8 @@ namespace Krypton.Toolkit
         // Custom
         private const string DEFAULT_COLLAPSE = "C&ollapse";
         private const string DEFAULT_EXPAND = "E&xpand";
+        private const string DEFAULT_MORE_DETAILS = "M&ore Details...";
+        private const string DEFAULT_LESS_DETAILS = "Les&s Details...";
         
         #endregion
 
@@ -85,7 +87,9 @@ namespace Krypton.Toolkit
                                  Continue.Equals(DEFAULT_CONTINUE) &&
                                  TryAgain.Equals(DEFAULT_TRY_AGAIN) &&
                                  Collapse.Equals(DEFAULT_COLLAPSE) &&
-                                 Expand.Equals(DEFAULT_EXPAND);
+                                 Expand.Equals(DEFAULT_EXPAND) &&
+                                 MoreDetails.Equals(DEFAULT_MORE_DETAILS) &&
+                                 LessDetails.Equals(DEFAULT_LESS_DETAILS);
 
         /// <summary>
         /// Reset all strings to default values.
@@ -240,6 +244,18 @@ namespace Krypton.Toolkit
         [Description(@"Expand string used in expandable footers.")]
         [DefaultValue(DEFAULT_EXPAND)]
         public string Expand { get; set; }
+
+        /// <summary>Gets or sets the more details string used in expandable footers.</summary>
+        [Localizable(true)]
+        [Category(@"Visuals")]
+        [Description(@"More details string used in expandable footers.")]
+        public string MoreDetails { get; set; }
+
+        /// <summary>Gets or sets the less details string used in expandable footers.</summary>
+        [Localizable(true)]
+        [Category(@"Visuals")]
+        [Description(@"Less details string used in expandable footers.")]
+        public string LessDetails { get; set; }
 #endregion
     }
 }